Renamed API map
